Added functionality to automatically update event states

+++ app/routes/event_options.py
@@ -3,10 +3,10 @@ from sanic_jwt import protected
 from sanic.request import Request
 from sanic.response import json
 from tortoise.transactions import atomic
-from app.db.models import EventOption, User, UserAndGroup, UserEventOptionResponse
+from app.db.models import EventOption, User, UserAndGroup, UserEventOptionResponse, Event
 from app.utils.tools import filter_dict_by_keys
 from app.utils.decorators import check_for_permission
-from app.utils.types import UserGroupPermissionEnum
+from app.utils.types import UserGroupPermissionEnum, EventStateEnum
 
 event_options = Blueprint("event_options", url_prefix="/event_options")
 
@@ -26,6 +26,10 @@ async def get_event_option(request: Request, my_user: User, event_option: EventO
 @atomic()
 async def update_event_option(request: Request, my_user: User, event_option: EventOption|None):
     if event_option:
+        await event_option.fetch_related("event")
+        if event_option.event.state == EventStateEnum.ARCHIVED:
+            return json({"error": f"The Event is Archived."}, status=403)
+
         await event_option.update_from_dict(filter_dict_by_keys(request.json,["date", "start_time", "end_time"]))
         await event_option.save()
         return json(event_option.to_dict())
@@ -39,6 +43,9 @@ async def update_event_option(request: Request, my_user: User, event_option: Eve
 @atomic()
 async def delete_event_option(request: Request, my_user: User, event_option: EventOption|None):
     if event_option:
+        await event_option.fetch_related("event")
+        if event_option.event.state == EventStateEnum.ARCHIVED:
+            return json({"error": f"The Event is Archived."}, status=403)
         await event_option.delete()
         return json({"message": f"Event Option deleted successfully"})
     else:
@@ -64,11 +71,14 @@ async def get_user_event_option_responses(request: Request, my_user: User, event
 @atomic()
 async def create_user_event_option_response(request: Request, my_user: User, event_option: EventOption|None):
     data = request.json
-    
+    await Event.update_state()
     if not event_option:
         return json({"error": f"EventOption not found"}, status=404)
     
     await event_option.fetch_related("event")
+    if event_option.event.state != EventStateEnum.VOTING:
+        return json({"error": f"The Voting period is over."}, status=403)
+    
     user_and_group = await UserAndGroup.get_or_none(user_id=my_user.id, group_id=event_option.event.group_id)
     if not user_and_group:
         return json({"error": f"User is not in Group"}, status=404)
@@ -92,6 +102,8 @@ async def create_user_event_option_response(request: Request, my_user: User, eve
 async def set_event_option_for_event(request: Request, my_user: User, event_option: EventOption|None):
     if event_option:
         await event_option.fetch_related("event")
+        if event_option.event.state == EventStateEnum.ARCHIVED:
+            return json({"error": f"The Event is Archived."}, status=403)
         event_option.event.choosen_event_option_id = event_option.id
         await event_option.event.save()
         return json({"message": f"Set EventOption for Event"})
